2025-2/RaizExemplo/requirements.txt
django
djangorestframework
django-cors-headers
setuptools
